build: enforce zero javac warnings (-Xlint:all -Werror)
Turn on every javac lint category and fail the build on any warning via
the compiler plugin (failOnWarning). Excludes only the classfile category,
which flags missing annotation class files inside third-party deps
(org.immutables.value.Generated, transitive) — unactionable in our code.

Clear the 13 pre-existing warnings to get to zero:
- serialVersionUID on VortexException, ProtoParseException
- drop redundant casts: PcoEncodingDecoder outer (int) (Math.clamp(long,int,int)
  already returns int; the load-bearing (long) on the subtraction stays),
  and (short)/(byte) on already-typed RAW in LazyConstantArrayTest
- @SuppressWarnings("unchecked") on JdbcImporter.fillCell (buffer type fixed
  by dtype at allocation)
- @SuppressWarnings("restricted") on the FFM downcall/lookup bindings in
  PosixTerminal and WindowsTerminal

+++ reader/src/test/java/io/github/dfa1/vortex/reader/array/LazyConstantArrayTest.java
@@ -128,7 +128,7 @@ class LazyConstantArrayTest {
         void fold_signedAndUnsignedWidening() {
             // signed short value is negative; unsigned is 0xFF00
             assertThat(new LazyConstantShortArray(new DType.Primitive(PType.I16, false), 2, RAW)
-                    .fold(0L, Long::sum)).isEqualTo(2L * (short) RAW);
+                    .fold(0L, Long::sum)).isEqualTo(2L * RAW);
             assertThat(new LazyConstantShortArray(new DType.Primitive(PType.U16, false), 2, RAW)
                     .fold(0L, Long::sum)).isEqualTo(2L * 0xFF00);
         }
@@ -140,7 +140,7 @@ class LazyConstantArrayTest {
 
             // When / Then — falls back to signed widening
             assertThat(sut.getInt(0)).isEqualTo((int) RAW);
-            assertThat(sut.fold(0L, Long::sum)).isEqualTo(2L * (short) RAW);
+            assertThat(sut.fold(0L, Long::sum)).isEqualTo(2L * RAW);
         }
 
         @Test
@@ -175,7 +175,7 @@ class LazyConstantArrayTest {
         @Test
         void fold_signedAndUnsignedWidening() {
             assertThat(new LazyConstantByteArray(new DType.Primitive(PType.I8, false), 2, RAW)
-                    .fold(0L, Long::sum)).isEqualTo(2L * (byte) RAW);
+                    .fold(0L, Long::sum)).isEqualTo(2L * RAW);
             assertThat(new LazyConstantByteArray(new DType.Primitive(PType.U8, false), 2, RAW)
                     .fold(0L, Long::sum)).isEqualTo(2L * 0xF0);
         }
@@ -187,7 +187,7 @@ class LazyConstantArrayTest {
 
             // When / Then — falls back to signed widening
             assertThat(sut.getInt(0)).isEqualTo((int) RAW);
-            assertThat(sut.fold(0L, Long::sum)).isEqualTo(2L * (byte) RAW);
+            assertThat(sut.fold(0L, Long::sum)).isEqualTo(2L * RAW);
         }
 
         @Test
